move APIFnModAgent to action_updateagent.go
